Offset based pagination
Pagination completed, all functionalities completed.
Pending - docker config file

+++ src/main/java/com/pranshu/crypto/models/dtos/PriceResponseDTO.java
@@ -8,6 +8,6 @@ import java.util.List;
 public class PriceResponseDTO {
     private String url; // current url
     private String next; // next url in pagination
-    private Integer count; // total no of records for the given query
+    private Long count; // total no of records for the given query
     private List<CoinPriceDTO> data; // list of price objects
 }
